backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
cohere>=4.0.0
qdrant-client>=1.7.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
pydantic>=2.5.0
numpy>=1.24.3
tiktoken>=0.5.2
beautifulsoup4>=4.12.0
requests>=2.31.0